LayoutTests/http/tests/security/local-image-from-remote.html

<html>
<head>
    <script>        
        if (window.layoutTestController)
            layoutTestController.dumpAsText();
        
        function imageTest() {
            var result = document.getElementById("result");

            var myImg = document.getElementById("myImg");
            if (myImg.height == 0 && myImg.width == 0) {
                result.innerHTML = "Test Passed.";
            } else {
                result.innerHTML = "Test Failed: Local image loaded remotely.";
            }
            
        }
    </script>
</head>
<body onLoad="imageTest()">
    <div id="div0">
        This test is to see if a remote file can include a local image.
    </div>
    </br>
    <div id="result">
        Test has not run.
    </div>
    </br>
    <img id="myImg" src="file:///tmp/LayoutTests/http/tests/security/resources/compass.jpg" />
</body>
</html>
