LayoutTests/http/tests/security/local-image-from-remote-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: file:///tmp/LayoutTests/http/tests/security/resources/compass.jpg
This test is to see if a remote file can include a local image.

Test Passed.
